+++ src/pages/PlansPage/components/PlanRow/index.ts
@@ -1,0 +1,1 @@
+export * from './PlanRow'
