+++ src/pages/PlansPage/components/PlanRow/style.module.scss
@@ -1,0 +1,4 @@
+.caseRow {
+    height: 60px;
+    align-content: center;
+}
