+++ src/pages/PlansPage/style.module.scss
@@ -9,11 +9,6 @@
     width: 90%;
 }
 
-.caseRow {
-    height: 60px;
-    align-content: center;
-}
-
 .tableContent {
     display: initial;
